During cohort fusion make cohort age the mean of the two fused cohorts weighted by cohort number in each. Make cohort age and cohort age fusion criterion real numbers - to be consistent with patch age.

+++ main/EDTypesMod.F90
@@ -152,7 +152,7 @@ module EDTypesMod
   ! COHORT FUSION
   real(r8), parameter :: HITEMAX              = 30.0_r8    ! max dbh value used in hgt profile comparison 
   integer , parameter :: N_HITE_BINS          = 60         ! no. of hite bins used to distribute LAI
-  integer, parameter  :: cohort_age_fusion_tol = 5.0_r8 * hlm_days_per_year  ! fusion tol based on cohort age
+  real(r8), parameter :: cohort_age_fusion_tol = 5.0_r8 * hlm_days_per_year  ! fusion tol based on cohort age
 
   ! COHORT TERMINATION
